+++ graphql/admin/admin.go
@@ -500,7 +500,7 @@ func (as *adminServer) addConnectedAdminResolvers() {
 		WithMutationResolver("addGroup",
 			func(m schema.Mutation) resolve.MutationResolver {
 				return resolve.NewMutationResolver(
-					resolve.NewAddGroupRewriter(),
+					NewAddGroupRewriter(),
 					resolve.DgraphAsQueryExecutor(),
 					resolve.DgraphAsMutationExecutor(),
 					resolve.StdMutationCompletion(m.Name()))
@@ -516,7 +516,7 @@ func (as *adminServer) addConnectedAdminResolvers() {
 		WithMutationResolver("updateGroup",
 			func(m schema.Mutation) resolve.MutationResolver {
 				return resolve.NewMutationResolver(
-					resolve.NewUpdateGroupRewriter(),
+					NewUpdateGroupRewriter(),
 					resolve.DgraphAsQueryExecutor(),
 					resolve.DgraphAsMutationExecutor(),
 					resolve.StdMutationCompletion(m.Name()))
