+++ graphql/resolve/mutation_rewriter.go
@@ -30,20 +30,18 @@ import (
 )
 
 const (
-	mutationQueryVar        = "x"
-	mutationQueryVarUID     = "uid(x)"
+	MutationQueryVar        = "x"
+	MutationQueryVarUID     = "uid(x)"
 	updateMutationCondition = `gt(len(x), 0)`
 )
 
-type addRewriter struct {
+type AddRewriter struct {
 	frags [][]*mutationFragment
 }
-type addGroupRewriter addRewriter
-type updateRewriter struct {
+type UpdateRewriter struct {
 	setFrags []*mutationFragment
 	delFrags []*mutationFragment
 }
-type updateGroupRewriter updateRewriter
 type deleteRewriter struct{}
 
 // A mutationFragment is a partially built Dgraph mutation.  Given a GraphQL
@@ -72,31 +70,23 @@ type mutationBuilder func(frag *mutationFragment) ([]byte, error)
 // result indicates that the upsert didn't succeed.
 type resultChecker func(map[string]interface{}) error
 
-// A variableGenerator generates unique variable names.
-type variableGenerator int
+// A VariableGenerator generates unique variable names.
+type VariableGenerator int
 
-// next gets the next variable name for the given type.
-func (c *variableGenerator) next(typ schema.Type) string {
+// Next gets the Next variable name for the given type.
+func (c *VariableGenerator) Next(typ schema.Type) string {
 	*c++
 	return fmt.Sprintf("%s%v", typ.Name(), int(*c))
 }
 
 // NewAddRewriter returns new MutationRewriter for add & update mutations.
 func NewAddRewriter() MutationRewriter {
-	return &addRewriter{}
-}
-
-func NewAddGroupRewriter() MutationRewriter {
-	return &addGroupRewriter{}
+	return &AddRewriter{}
 }
 
 // NewUpdateRewriter returns new MutationRewriter for add & update mutations.
 func NewUpdateRewriter() MutationRewriter {
-	return &updateRewriter{}
-}
-
-func NewUpdateGroupRewriter() MutationRewriter {
-	return &updateGroupRewriter{}
+	return &UpdateRewriter{}
 }
 
 // NewDeleteRewriter returns new MutationRewriter for delete mutations..
@@ -181,7 +171,7 @@ func NewDeleteRewriter() MutationRewriter {
 //   } ],
 //   "Author.friends":[ {"uid":"0x123"} ],
 // }
-func (mrw *addRewriter) Rewrite(
+func (mrw *AddRewriter) Rewrite(
 	m schema.Mutation) (*gql.GraphQuery, []*dgoapi.Mutation, error) {
 
 	mutatedType := m.MutatedType()
@@ -190,7 +180,7 @@ func (mrw *addRewriter) Rewrite(
 		return mrw.handleMultipleMutations(m)
 	}
 
-	varGen := variableGenerator(0)
+	varGen := VariableGenerator(0)
 	val := m.ArgValue(schema.InputArgName).(map[string]interface{})
 	mrw.frags = [][]*mutationFragment{rewriteObject(mutatedType, nil, "", &varGen, true, val)}
 	mutations, err := mutationsFromFragments(
@@ -210,12 +200,12 @@ func (mrw *addRewriter) Rewrite(
 		schema.GQLWrapf(err, "failed to rewrite mutation payload")
 }
 
-func (mrw *addRewriter) handleMultipleMutations(
+func (mrw *AddRewriter) handleMultipleMutations(
 	m schema.Mutation) (*gql.GraphQuery, []*dgoapi.Mutation, error) {
 	mutatedType := m.MutatedType()
 	val, _ := m.ArgValue(schema.InputArgName).([]interface{})
 
-	varGen := variableGenerator(0)
+	varGen := VariableGenerator(0)
 	var errs error
 	var mutationsAll []*dgoapi.Mutation
 	queries := &gql.GraphQuery{}
@@ -255,7 +245,7 @@ func (mrw *addRewriter) handleMultipleMutations(
 }
 
 // FromMutationResult rewrites the query part of a GraphQL add mutation into a Dgraph query.
-func (mrw *addRewriter) FromMutationResult(
+func (mrw *AddRewriter) FromMutationResult(
 	mutation schema.Mutation,
 	assigned map[string]string,
 	result map[string]interface{}) (*gql.GraphQuery, error) {
@@ -295,34 +285,6 @@ func (mrw *addRewriter) FromMutationResult(
 	return rewriteAsQueryByIds(mutation.QueryField(), uids), errs
 }
 
-// Rewrite rewrites schema.Mutation into GraphQL+- upsert mutations only for Group type.
-// It ensures that only the last rule out of all duplicate rules in input is preserved.
-// A rule is duplicate if it has same predicate name as another rule.
-func (mrw *addGroupRewriter) Rewrite(
-	m schema.Mutation) (*gql.GraphQuery, []*dgoapi.Mutation, error) {
-	addGroupInput, _ := m.ArgValue(schema.InputArgName).([]interface{})
-
-	// remove rules with same predicate name for each group input
-	for i, groupInput := range addGroupInput {
-		rules, _ := groupInput.(map[string]interface{})["rules"].([]interface{})
-		rules = removeDuplicateRuleRef(rules)
-		addGroupInput[i].(map[string]interface{})["rules"] = rules
-	}
-
-	m.SetArgTo(schema.InputArgName, addGroupInput)
-
-	return ((*addRewriter)(mrw)).Rewrite(m)
-}
-
-// FromMutationResult rewrites the query part of a GraphQL add mutation into a Dgraph query.
-func (mrw *addGroupRewriter) FromMutationResult(
-	mutation schema.Mutation,
-	assigned map[string]string,
-	result map[string]interface{}) (*gql.GraphQuery, error) {
-
-	return ((*addRewriter)(mrw)).FromMutationResult(mutation, assigned, result)
-}
-
 // Rewrite rewrites set and remove update patches into GraphQL+- upsert mutations.
 // The GraphQL updates look like:
 //
@@ -344,8 +306,8 @@ func (mrw *addGroupRewriter) FromMutationResult(
 // - Explicit values in remove mean delete this if it is the actual value
 // - Nulls in remove become like delete * for the corresponding predicate.
 //
-// See addRewriter for how the set and remove fragments get created.
-func (urw *updateRewriter) Rewrite(
+// See AddRewriter for how the set and remove fragments get created.
+func (urw *UpdateRewriter) Rewrite(
 	m schema.Mutation) (*gql.GraphQuery, []*dgoapi.Mutation, error) {
 
 	mutatedType := m.MutatedType()
@@ -358,12 +320,12 @@ func (urw *updateRewriter) Rewrite(
 		return nil, nil, nil
 	}
 
-	upsertQuery := rewriteUpsertQueryFromMutation(m)
-	srcUID := mutationQueryVarUID
+	upsertQuery := RewriteUpsertQueryFromMutation(m)
+	srcUID := MutationQueryVarUID
 
 	var errSet, errDel error
 	var mutSet, mutDel []*dgoapi.Mutation
-	varGen := variableGenerator(0)
+	varGen := VariableGenerator(0)
 
 	if setArg != nil {
 		urw.setFrags =
@@ -414,7 +376,7 @@ func (urw *updateRewriter) Rewrite(
 }
 
 // FromMutationResult rewrites the query part of a GraphQL update mutation into a Dgraph query.
-func (urw *updateRewriter) FromMutationResult(
+func (urw *UpdateRewriter) FromMutationResult(
 	mutation schema.Mutation,
 	assigned map[string]string,
 	result map[string]interface{}) (*gql.GraphQuery, error) {
@@ -445,170 +407,6 @@ func (urw *updateRewriter) FromMutationResult(
 	}
 
 	return rewriteAsQueryByIds(mutation.QueryField(), uids), nil
-}
-
-// Rewrite rewrites set and remove update patches into GraphQL+- upsert mutations
-// only for Group type. It ensures that if a rule already exists in db, it is updated;
-// otherwise, it is created. It also ensures that only the last rule out of all
-// duplicate rules in input is preserved. A rule is duplicate if it has same predicate
-// name as another rule.
-func (urw *updateGroupRewriter) Rewrite(m schema.Mutation) (*gql.GraphQuery,
-	[]*dgoapi.Mutation, error) {
-	inp := m.ArgValue(schema.InputArgName).(map[string]interface{})
-	setArg := inp["set"]
-	delArg := inp["remove"]
-
-	if setArg == nil && delArg == nil {
-		return nil, nil, nil
-	}
-
-	upsertQuery := rewriteUpsertQueryFromMutation(m)
-	srcUID := mutationQueryVarUID
-
-	var mutSet, mutDel []*dgoapi.Mutation
-	queries := []*gql.GraphQuery{upsertQuery}
-	varGen := variableGenerator(0)
-	ruleType := m.MutatedType().Field("rules").Type()
-
-	if setArg != nil {
-		rules, _ := setArg.(map[string]interface{})["rules"].([]interface{})
-		rules = removeDuplicateRuleRef(rules)
-		for _, ruleI := range rules {
-			rule := ruleI.(map[string]interface{})
-			variable := varGen.next(ruleType)
-			predicate := rule["predicate"]
-			permission := rule["permission"]
-
-			ruleQuery := writeAclRuleQuery(m, predicate.(string), variable)
-			queries = append(queries, ruleQuery)
-
-			nonExistentJson := []byte(fmt.Sprintf(`
-			{
-				"uid": "%s",
-				"dgraph.acl.rule": [
-					{
-						"uid":                    "_:%s",
-						"dgraph.type":            "%s",
-						"dgraph.rule.predicate":  "%s",
-						"dgraph.rule.permission": %v
-					}
-				]
-			}`, srcUID, variable, ruleType.DgraphName(), predicate, permission))
-
-			existsJson := []byte(fmt.Sprintf(`
-			{
-				"uid":                    "uid(%s)",
-				"dgraph.rule.permission": %v
-			}`, variable, permission))
-
-			mutSet = append(mutSet, &dgoapi.Mutation{
-				SetJson: nonExistentJson,
-				Cond: fmt.Sprintf(`@if(gt(len(%s),0) AND eq(len(%s),0))`, mutationQueryVar,
-					variable),
-			}, &dgoapi.Mutation{
-				SetJson: existsJson,
-				Cond: fmt.Sprintf(`@if(gt(len(%s),0) AND gt(len(%s),0))`, mutationQueryVar,
-					variable),
-			})
-		}
-	}
-
-	if delArg != nil {
-		rules, _ := delArg.(map[string]interface{})["rules"].([]interface{})
-		for _, predicate := range rules {
-			variable := varGen.next(ruleType)
-
-			ruleQuery := writeAclRuleQuery(m, predicate.(string), variable)
-			queries = append(queries, ruleQuery)
-
-			deleteJson := []byte(fmt.Sprintf(`
-			{
-				"uid": "%s",
-				"dgraph.acl.rule": [
-					{
-						"uid":                    "uid(%s)",
-						"dgraph.type":            "",
-						"dgraph.rule.predicate":  "",
-						"dgraph.rule.permission": ""
-					}
-				]
-			}`, srcUID, variable))
-
-			mutDel = append(mutDel, &dgoapi.Mutation{
-				DeleteJson: deleteJson,
-				Cond: fmt.Sprintf(`@if(gt(len(%s),0) AND gt(len(%s),0))`, mutationQueryVar,
-					variable),
-			})
-		}
-	}
-
-	// if there is no mutation being performed as a result of some specific input,
-	// then we don't need to do the upsertQuery for group
-	if len(mutSet) == 0 && len(mutDel) == 0 {
-		return nil, nil, nil
-	}
-
-	return &gql.GraphQuery{Children: queries}, append(mutSet, mutDel...), nil
-}
-
-// FromMutationResult rewrites the query part of a GraphQL update mutation into a Dgraph query.
-func (urw *updateGroupRewriter) FromMutationResult(
-	mutation schema.Mutation,
-	assigned map[string]string,
-	result map[string]interface{}) (*gql.GraphQuery, error) {
-
-	return ((*updateRewriter)(urw)).FromMutationResult(mutation, assigned, result)
-}
-
-// writeAclRuleQuery returns a *gql.GraphQuery to query a rule inside a group based on
-// its predicate value.
-func writeAclRuleQuery(m schema.Mutation, predicate, variable string) *gql.GraphQuery {
-	ruleQuery := rewriteUpsertQueryFromMutation(m)
-	ruleQuery.Attr = variable
-	ruleQuery.Var = ""
-	ruleQuery.Children = append(ruleQuery.Children, &gql.GraphQuery{
-		Attr: "dgraph.acl.rule",
-		Var:  variable,
-		Filter: &gql.FilterTree{
-			Op:    "",
-			Child: nil,
-			Func: &gql.Function{
-				Name: "eq",
-				Args: []gql.Arg{
-					{
-						Value: "dgraph.rule.predicate",
-					},
-					{
-						Value: predicate,
-					},
-				},
-			},
-		},
-	})
-
-	return ruleQuery
-}
-
-// removeDuplicateRuleRef removes duplicate rules based on predicate value.
-// for duplicate rules, only the last rule with duplicate predicate name is preserved.
-func removeDuplicateRuleRef(rules []interface{}) []interface{} {
-	predicateMap := make(map[string]int, len(rules))
-	i := 0
-
-	for _, rule := range rules {
-		predicate, _ := rule.(map[string]interface{})["predicate"].(string)
-
-		// this ensures that only the last rule with duplicate predicate name is preserved
-		if idx, ok := predicateMap[predicate]; !ok {
-			predicateMap[predicate] = i
-			rules[i] = rule
-			i++
-		} else {
-			rules[idx] = rule
-		}
-	}
-
-	return rules[:i]
 }
 
 func extractMutated(result map[string]interface{}, mutatedField string) []string {
@@ -671,10 +469,10 @@ func extractFilter(m schema.Mutation) map[string]interface{} {
 	return filter
 }
 
-func rewriteUpsertQueryFromMutation(m schema.Mutation) *gql.GraphQuery {
+func RewriteUpsertQueryFromMutation(m schema.Mutation) *gql.GraphQuery {
 	// The query needs to assign the results to a variable, so that the mutation can use them.
 	dgQuery := &gql.GraphQuery{
-		Var:  mutationQueryVar,
+		Var:  MutationQueryVar,
 		Attr: m.ResponseName(),
 	}
 	// Add uid child to the upsert query, so that we can get the list of nodes upserted.
@@ -702,8 +500,8 @@ func (drw *deleteRewriter) Rewrite(m schema.Mutation) (
 			m.MutationType())
 	}
 
-	varGen := variableGenerator(0)
-	qry := rewriteUpsertQueryFromMutation(m)
+	varGen := VariableGenerator(0)
+	qry := RewriteUpsertQueryFromMutation(m)
 	deletes := []interface{}{map[string]interface{}{"uid": "uid(x)"}}
 
 	// we need to delete this node with ^^ and then any reference we know about
@@ -718,7 +516,7 @@ func (drw *deleteRewriter) Rewrite(m schema.Mutation) (
 				continue
 			}
 		}
-		varName := varGen.next(fld.Type())
+		varName := varGen.Next(fld.Type())
 
 		qry.Children = append(qry.Children,
 			&gql.GraphQuery{
@@ -727,7 +525,7 @@ func (drw *deleteRewriter) Rewrite(m schema.Mutation) (
 			})
 
 		delFldName := fld.Type().DgraphPredicate(invField.Name())
-		del := map[string]interface{}{"uid": mutationQueryVarUID}
+		del := map[string]interface{}{"uid": MutationQueryVarUID}
 		if invField.Type().ListType() == nil {
 			deletes = append(deletes,
 				map[string]interface{}{
@@ -832,7 +630,7 @@ func queryFromFragments(frags []*mutationFragment) *gql.GraphQuery {
 	return qry
 }
 
-// rewriteObject rewrites obj to a list of mutation fragments.  See addRewriter.Rewrite
+// rewriteObject rewrites obj to a list of mutation fragments.  See AddRewriter.Rewrite
 // for a description of what those fragments look like.
 //
 // GraphQL validation has already ensured that the types of arguments (or variables)
@@ -851,14 +649,14 @@ func rewriteObject(
 	typ schema.Type,
 	srcField schema.FieldDefinition,
 	srcUID string,
-	varGen *variableGenerator,
+	varGen *VariableGenerator,
 	withAdditionalDeletes bool,
 	obj map[string]interface{}) []*mutationFragment {
 
 	atTopLevel := srcField == nil
 	topLevelAdd := srcUID == ""
 
-	variable := varGen.next(typ)
+	variable := varGen.Next(typ)
 
 	id := typ.IDField()
 	if id != nil {
@@ -1063,7 +861,7 @@ func asIDReference(
 	srcUID string,
 	variable string,
 	withAdditionalDeletes bool,
-	varGen *variableGenerator) *mutationFragment {
+	varGen *VariableGenerator) *mutationFragment {
 
 	result := make(map[string]interface{}, 2)
 	frag := newFragment(result)
@@ -1125,7 +923,7 @@ func asXIDReference(
 	typ schema.Type,
 	xidFieldName, xidString, xidVariable string,
 	withAdditionalDeletes bool,
-	varGen *variableGenerator) *mutationFragment {
+	varGen *VariableGenerator) *mutationFragment {
 
 	result := make(map[string]interface{}, 2)
 	frag := newFragment(result)
@@ -1165,7 +963,7 @@ func asXIDReference(
 // leave an edge.
 func addAdditionalDeletes(
 	frag *mutationFragment,
-	varGen *variableGenerator,
+	varGen *VariableGenerator,
 	srcField schema.FieldDefinition, srcUID, variable string) {
 
 	if srcField == nil {
@@ -1182,7 +980,7 @@ func addAdditionalDeletes(
 }
 
 func addDelete(frag *mutationFragment,
-	varGen *variableGenerator,
+	varGen *VariableGenerator,
 	qryVar, excludeVar string,
 	qryFld, delFld schema.FieldDefinition) {
 
@@ -1199,7 +997,7 @@ func addDelete(frag *mutationFragment,
 		qryVar = qryVar[4 : len(qryVar)-1]
 	}
 
-	targetVar := varGen.next(qryFld.Type())
+	targetVar := varGen.Next(qryFld.Type())
 	delFldName := qryFld.Type().DgraphPredicate(delFld.Name())
 
 	qry := &gql.GraphQuery{
@@ -1283,7 +1081,7 @@ func rewriteList(
 	typ schema.Type,
 	srcField schema.FieldDefinition,
 	srcUID string,
-	varGen *variableGenerator,
+	varGen *VariableGenerator,
 	withAdditionalDeletes bool,
 	objects []interface{}) []*mutationFragment {
